aerodatabox_proxy/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
requests>=2.28.0
python-dotenv>=1.0.0
pydantic>=1.10.0
starlette>=0.27.0
python-dateutil>=2.8.2
pytz>=2023.3
itsdangerous>=2.1.2
jinja2>=3.1.2
markupsafe>=2.1.2